refactor Analog Clock
